+++ mscv-resenia/src/main/java/com/smedinamsvc/resenia/service/ReseniaService.java
@@ -7,9 +7,9 @@ import com.smedinamsvc.resenia.model.Resenia;
 public interface ReseniaService {
 
     List<Resenia> findAll();                              // Lista todas las reseñas
-    List<Resenia> findByProductoId(Integer productoId);   // Lista todas las reseñas de un producto
-    Resenia findById(Integer id);                         // Busca reseña por su id
+    List<Resenia> findByProductoId(Long productoId);   // Lista todas las reseñas de un producto
+    Resenia findById(Long id);                         // Busca reseña por su id
     Resenia save(Resenia resenia);                        // Guarda una reseña completa (objeto)
-    void deleteById(Integer id);                          // Elimina reseña por id (void)
+    void deleteById(Long id);                          // Elimina reseña por id (void)
 
 }
